Move framework to pkg
Pkg is a more Go-esque name for a framework directory and plays nicely with building the framework alongside a working application. The routing files were moved to this location.
