--- framework/README.md
@@ -1,2 +1,0 @@
-# Framework package
-A place to test ideas in-place before eventually promoting them to a proper framework.
