+++ pkg/README.md
@@ -1,0 +1,3 @@
+# Framework Packages
+
+Where all the various package sources are stored.
